requirements.txt
amply==0.1.4
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
docutils==0.17
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
PuLP==2.4
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
twilio==6.58.0
urllib3==1.26.4
